requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.4.0
django-filter==23.3
markdown==3.5.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7